Move IntVector2 into Geometry namespace
And move IntVector2Extensions to Geometry.Extensions namespace
