+++ Assets/Scripts/Clipboard/Clipboard.cs
@@ -1,8 +1,8 @@
 using PAC.Animation;
 using PAC.Colour.Compositing;
-using PAC.DataStructures;
 using PAC.Drawing;
 using PAC.Files;
+using PAC.Geometry;
 using PAC.ImageEditing;
 using PAC.Layers;
