+++ Assets/Scripts/Colour/Compositing/BlendMode.cs
@@ -2,8 +2,8 @@ using System;
 using System.Linq;
 using System.Reflection;
 
-using PAC.DataStructures;
 using PAC.Extensions;
+using PAC.Geometry;
 using PAC.Json;
 
 using UnityEngine;
