+++ Assets/Scripts/Drawing/DrawingArea.cs
@@ -20,7 +20,8 @@ using PAC.ImageEditing;
 using PAC.Colour.Compositing;
 using PAC.Geometry.Shapes.Extensions;
 using PAC.Patterns.Extensions;
-using PAC.DataStructures.Extensions;
+using PAC.Geometry;
+using PAC.Geometry.Extensions;
 
 namespace PAC.Drawing
 {
