+++ Assets/Scripts/Geometry/Shapes/Extensions/IShapeExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using PAC.DataStructures;
-using PAC.DataStructures.Extensions;
 using PAC.Extensions;
+using PAC.Geometry.Extensions;
 using PAC.Geometry.Shapes.Interfaces;
 using PAC.ImageEditing;
