+++ Assets/Scripts/Image Editing/Outline.cs
@@ -1,10 +1,10 @@
 using System;
 using System.Collections.Generic;
 
-using PAC.DataStructures;
 using PAC.DataStructures.Extensions;
 using PAC.Extensions;
 using PAC.Geometry;
+using PAC.Geometry.Extensions;
 
 using UnityEngine;
