+++ Assets/Scripts/Patterns/Extensions/IPattern2DExtensions.cs
@@ -2,6 +2,7 @@ using System;
 
 using PAC.DataStructures;
 using PAC.Extensions;
+using PAC.Geometry;
 
 using UnityEngine;
