+++ Assets/Scripts/Patterns/GradientLinear.cs
@@ -1,4 +1,4 @@
-using PAC.DataStructures;
+using PAC.Geometry;
 
 using UnityEngine;
