+++ Assets/Scripts/Screen/ScreenInfo.cs
@@ -1,4 +1,5 @@
-using PAC.DataStructures;
+using PAC.Geometry;
+
 using UnityEngine;
 
 namespace PAC.Screen
