+++ Assets/Scripts/Tilesets/Tile.cs
@@ -1,8 +1,11 @@
 using System.Collections.Generic;
 using System.Linq;
+
 using PAC.DataStructures;
 using PAC.Files;
+using PAC.Geometry;
 using PAC.Layers;
+
 using UnityEngine;
 using UnityEngine.Events;
