+++ Assets/Tests/Geometry/IntVector2_Tests.cs
@@ -1,6 +1,6 @@
 using NUnit.Framework;
 
-using PAC.DataStructures;
+using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Maths;
