+++ Assets/Tests/Geometry/Shapes/Default Tests/IFlippableShape_DefaultTests.cs
@@ -3,8 +3,8 @@ using System.Linq;
 
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Exceptions;
+using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes.Interfaces;
 using PAC.Tests.Geometry.Shapes.RequiredTests;
