+++ Assets/Tests/Geometry/Shapes/Default Tests/IShape_DefaultTests.cs
@@ -6,6 +6,7 @@ using NUnit.Framework;
 
 using PAC.DataStructures;
 using PAC.Exceptions;
+using PAC.Geometry;
 using PAC.Geometry.Shapes.Interfaces;
 using PAC.Tests.Geometry.Shapes.RequiredTests;
 using PAC.Tests.Geometry.Shapes.TestUtils;
