+++ Assets/Tests/Geometry/Shapes/Default Tests/ITranslatableShape_DefaultTests.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using NUnit.Framework;
 
 using PAC.DataStructures;
+using PAC.Geometry;
 using PAC.Geometry.Shapes.Interfaces;
 using PAC.Tests.Geometry.Shapes.RequiredTests;
 using PAC.Tests.Geometry.Shapes.TestUtils;
