+++ Assets/Tests/Geometry/Shapes/Rectangle_Tests.cs
@@ -2,6 +2,7 @@ using NUnit.Framework;
 
 using PAC.DataStructures;
 using PAC.Extensions;
+using PAC.Geometry;
 using PAC.Geometry.Shapes;
 using PAC.Tests.Geometry.Shapes.DefaultTests;
 using PAC.Tests.Geometry.Shapes.RequiredTests;
