+++ Assets/Tests/Geometry/Shapes/TestUtils/ShapeUtils.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 
-using PAC.DataStructures;
+using PAC.Geometry;
 
 namespace PAC.Tests.Geometry.Shapes.TestUtils
 {
