+++ Assets/Tests/Patterns/Checkerboard_Tests.cs
@@ -1,6 +1,7 @@
 using NUnit.Framework;
 
 using PAC.DataStructures;
+using PAC.Geometry;
 using PAC.Patterns;
 
 using UnityEngine;
