+++ Assets/Tests/Patterns/GradientLinear_Tests.cs
@@ -5,6 +5,7 @@ using NUnit.Framework;
 
 using PAC.DataStructures;
 using PAC.Extensions;
+using PAC.Geometry;
 
 using UnityEngine;
